# Improve performance when using `--long`