@solidiquis 2024-03-13
Amazing thanks! I'll be sure to include this in v4 which I'm currently working on.